zeroshot-clip/requirements.txt
torch
torchvision
Pillow
matplotlib
tqdm
dataclasses
openai-clip
opencv-python
numpy
scikit-learn